Fix system ordering and add game over test
System ordering turns out to be a critical part of 2048 logic. Specifically
the most relevant ordering is:

1. board shift
2. new tile addition

I've added chain to more than is necessary here, but since it is in no
way performance sensitive, it might be easier than running "as much as possible"
in parallel.

This commit also adds an example testing the game over system and adds
some cfg(test) flagged functionality to support that testing.