Spawning a 2d Camera
If we want to see anything rendered on screen, we need to set up a `Camera`. To do that we'll bootstrap a new [Camera2dBundle](https://docs.rs/bevy/0.13.2/bevy/prelude/struct.Camera2dBundle.html) using the defaults since 2048 is a 2d game.

```rust
fn setup(mut commands: Commands) {
    commands.spawn(Camera2dBundle::default());
}
```

In addition to the concept of a Camera, there are three new ideas here: Bundles, Commands, and spawning entities.

\## Cameras

A Camera is an entity in Bevy with components attached to it. You can think of the Camera as almost exactly the same thing as a real DSLR digital camera or the camera on your iPhone. The image you'd get out of a real camera is the image you're rendering to the screen in Bevy. A camera is responsible for determining quite a few aspects for how our app is rendered including fancy things like the view projection. For us in this project all of the defaults work just fine.

\## Bundles

Bundles are collections of components. They make it a bit easier to add lots of components at one time. For example, the `Camera2dBundle` includes about 10 components we didn't have to set up or configure because we used the `default()` method which gave us a `Bundle` of components that set up a 2d camera for us.

\## Commands and Spawning

In a Bevy game, basically everything is accessible through the [World](https://docs.rs/bevy/0.13.2/bevy/ecs/prelude/struct.World.html) struct even though we mostly don't see it. The [World](https://docs.rs/bevy/0.13.2/bevy/ecs/prelude/struct.World.html) struct stores and exposes operations on entities, components, and their associated metadata. `Commands` are how we perform structural changes to the World. In this example we use a command called [spawn](https://docs.rs/bevy/0.13.2/bevy/ecs/system/struct.Commands.html#method.spawn) to insert a new camera entity with the appropriate components into our game.

\## Adding Systems

Now that we have a `setup` function, we can use that as a system in our Bevy `App`. We only want the setup function to run once at startup because we only ever need one camera for our game. To do this we can add our `startup` system to the [`Startup`](https://docs.rs/bevy/0.13.2/bevy/app/struct.Main.html) schedule using [`add_systems`](https://docs.rs/bevy/0.13.2/bevy/app/struct.App.html#method.add_systems) on the [App](https://docs.rs/bevy/0.13.2/bevy/app/struct.App.html).

```rust
use bevy::prelude::*;

fn main() {
    App::new()
        .add_plugins(DefaultPlugins.set(WindowPlugin {
            primary_window: Some(Window {
                title: "2048".to_string(),
                ..default()
            }),
            ..default()
        }))
        .add_systems(Startup, setup)
        .run()
}

fn setup(mut commands: Commands) {
    commands.spawn(Camera2dBundle::default());
}
```

We can pass a function (or many functions) into `add_systems` and Bevy will turn our function into a system for us.

Bevy turning out function into a system is how our function gets it's `Commands` argument. Bevy understands from the type signature what our system needs, and makes sure that is available when we want to run it. We will use this functionality later to do more interesting queries for our systems.

If we `cargo run` now, we'll see that the background color of our app changed to a gray color. This is called the [`ClearColor`](https://docs.rs/bevy/0.13.2/bevy/prelude/struct.ClearColor.html) and it is used when there's nothing else for a camera to render. This is how we know the camera is working.

![A blank Bevy window with a Camera](https://res.cloudinary.com/dilgcuzda/image/upload/v1713908141/workshops/2048-with-bevy-ecs/bevy-0.13/004-window-with-camera.avif)