Creating a Board using Bevy's Resources
In this lesson we're going to create the board. This will involve creating a `Resource` representing our `Board` as well as spawning a `Sprite` so we can see the `Board` on screen.

The `Board` is going to be a regular struct that contains one field, `size`, that indicates how big the `Board` is going to be.

```rust
\#[derive(Resource)]
struct Board {
    size: u8,
}
```

`Resource` is both a [trait](https://docs.rs/bevy/0.13.2/bevy/ecs/system/trait.Resource.html) and a [derive macro](https://docs.rs/bevy/0.13.2/bevy/ecs/system/derive.Resource.html). We derive `Resource` on the `Board` so that the Rust compiler will handle writing the implementation of the `Resource` trait for `Board`.

A `Resource` is singleton, global data. This means we can have exactly one `Board` and we can access that `Board` in our systems.

We'll insert a new `Board` using the `App` builder functions. Our board will be a 4x4 grid of tiles.

```rust
fn main() {
    App::new()
        .insert_resource(Board { size: 4 })
        .add_plugins(DefaultPlugins.set(WindowPlugin {
            primary_window: Some(Window {
                title: "2048".to_string(),
                ..default()
            }),
            ..default()
        }))
        .add_systems(Startup, setup)
        .run()
}
```

\## Spawning a Board

To place a board on the screen, we'll add a new startup system to handle spawning it. This looks very similar to the `setup` system we added to handle the camera.

`add_systems` accepts a tuple of multiple functions, so we can add the new `spawn_board` system there.

```rust
fn main() {
    App::new()
        .insert_resource(Board { size: 4 })
        .add_plugins(DefaultPlugins.set(WindowPlugin {
            primary_window: Some(Window {
                title: "2048".to_string(),
                ..default()
            }),
            ..default()
        }))
        .add_systems(Startup, (setup, spawn_board))
        .run()
}
```

We'll create a function called `spawn_board` that takes a mutable argument of type `Commands` so that we can create the entity for our board. The [type signature](https://docs.rs/bevy/0.13.2/bevy/ecs/system/struct.Commands.html#method.spawn) for `spawn` indicates that it must be able to mutate `commands`, so the `commands` argument must be specified as `mut`.

We also use a `const` here intentionally so that we don't have any magic, unlabelled numbers. You can think of the number we're using here here as number of pixels for now.

```
const TILE_SIZE: f32 = 40.0;

fn spawn_board(mut commands: Commands, board: Res<Board>) {
    let physical_board_size =
        f32::from(board.size) * TILE_SIZE;

    commands.spawn(SpriteBundle {
        sprite: Sprite {
            custom_size: Some(Vec2::splat(
                physical_board_size,
            )),
            ..default()
        },
        ..default()
    });
}
```

We also ask Bevy for the `Board` resource so that we know how big the board sprite needs to be.

First we use the `Board`'s `size` field and an abitrary `TILE_SIZE` to determine how big the on-screen sprite should be. I've termed this on-screen size the `physical_board_size` but there's nothing special about this terminology.

The `size` field is a [u8](https://doc.rust-lang.org/std/primitive.u8.html). Our board never gets bigger than 4x4, so a [u8](https://doc.rust-lang.org/std/primitive.u8.html) (which is an unsigned integer that can hold a maximum value of 255) is fine for our use case.

While our board is a 4x4 grid and we can represent that with integers, when we try to render that our into the game we have to translate the [u8](https://doc.rust-lang.org/std/primitive.u8.html) into an [f32](https://doc.rust-lang.org/std/primitive.f32.html). For various reasons, [f32](https://doc.rust-lang.org/std/primitive.f32.html) is a representation of numbers that Bevy likes so we often will have to convert from what number types we have into [f32s](https://doc.rust-lang.org/std/primitive.f32.html).

The `physical_board_size` here is the `board.size` (4) multiplied by the `TILE_SIZE`. This gives us how big the tile should be when rendered to the screen, almost like how many pixels it should occupy. If the board is a 4x4 grid of squares and each square is 10 pixels wide, then the board is 40px by 40px total.

We can convert from `u8` to `f32` using `f32::from()`. `u8` fits neatly into an `f32` and will never fail, so using `from` is appropriate here as it is guaranteed to never fail.

`commands.spawn` creates an entity for us and attaches some sprite related `Component`s to that entity using a [`SpriteBundle`](https://docs.rs/bevy/0.13.2/bevy/prelude/struct.SpriteBundle.html). The `SpriteBundle` contains a whole set of `Component`s that represent what a sprite can do, such as `Sprite`, `Transform`, and `Visibility`.

```rust
SpriteBundle {
    sprite: Sprite {
        custom_size: Some(Vec2::splat(
            physical_board_size,
        )),
        ..default()
    },
    ..default()
}
```

The `SpriteBundle` is constructed by filling out the various fields in the struct. One of these is `sprite`, which we care about because it is how we'll specify the size of our [`Sprite`](https://docs.rs/bevy/0.13.2/bevy/prelude/struct.Sprite.html). All of the rest of the fields can be their default values.

`SpriteBundle` implements the `Default` trait, which is a standard Rust trait that anyone can implement. This means we can use [struct update syntax](https://doc.rust-lang.org/stable/book/ch05-01-defining-structs.html#creating-instances-from-other-instances-with-struct-update-syntax), which is basically the JavaScript spread operator for objects. Note that there are only two dots, instead of three like in JavaScript.

The [`default()`](https://docs.rs/bevy/0.13.2/bevy/prelude/fn.default.html) method is a convenience function provided by Bevy's prelude which will call out to `SpriteBundle`'s and `Sprite`'s implementation of the `Default` trait to fill out the rest of the fields for us.

With the rest of the fields handled, we can focus on creating a new `Sprite`. The [`Sprite` struct](https://docs.rs/bevy/0.13.2/bevy/prelude/struct.Sprite.html) lets us specify the size as a `Vec2`, which we can create using the `physical_board_size` since our board is a square.

```rust
Sprite {
    custom_size: Some(Vec2::splat(
        physical_board_size,
    )),
    ..default()
}
```

Bevy provides a couple different `Vec*` variants. Since we're building a 2d game we'll be working with `Vec2`, a 2-dimensional vector that includes `x` and `y` values. There are also `Vec3` and `Vec4` for 3-dimensional and 4-dimensional vectors.

Rust has a `Vec` type already, so how are Bevy's different? The `Vec` type in Rust is like an array in JavaScript; a "contiguous growable array type" if that makes sense to you. You can add arbitrarily many items to it and it'll just keep working and growing. Bevy's `Vec2`, `Vec3`, and `Vec4` are similar, but restricted to 2, 3 and 4 items in them and optimized for those use cases. These items are often coordinates like x, y, and z.

sidenote: Bevy uses the [glam](https://crates.io/crates/glam) crate to power its Vec* types.

Running `cargo run` at this point should show a white box in the middle of the screen.

![a board on screen](https://res.cloudinary.com/dilgcuzda/image/upload/v1713937997/workshops/2048-with-bevy-ecs/bevy-0.13/005-board.avif)