Changing the color of the background with ClearColor
With a fresh new coat of paint for the board we'll change the background as well.

Our camera is responsible for painting all of the pixels we see on screen. Any area of the screen that doesn't have a sprite on it will be painted using a “clear color” (which is named like that because its the color used to clear the screen in between frames).

We can configure what the [`ClearColor`](https://docs.rs/bevy/0.13.2/bevy/prelude/struct.ClearColor.html) is by using a [`Resource`](https://docs.rs/bevy/0.13.2/bevy/ecs/prelude/trait.Resource.html) which will get accessed and used by Bevy's rendering system.

We created our own `Resource` earlier when we created our `Board` and we'll do the same here with the pre-existing `ClearColor`.

```rust
fn main() {
    App::new()
        .insert_resource(ClearColor(
            Color::hex("#1f2638")
              .expect("developer should have provided a valid hex code")
        ))
        .insert_resource(Board { size: 4 })
        .add_plugins(DefaultPlugins.set(WindowPlugin {
            primary_window: Some(Window {
                title: "2048".to_string(),
                ..default()
            }),
            ..default()
        }))
        .add_systems(Startup, (setup, spawn_board))
        .run()
}
```

This time we'll use the [`hex`](https://docs.rs/bevy/0.13.2/bevy/render/color/enum.Color.html#method.hex) function on `Color` to parse a hex code into a Color. Because this is parsing: it could fail if we don't give it a well formatted hex code, which is why the `hex` function returns a `Result<Color, HexColorError>`.

If we use a hex code that doesn't parse we'd rather see our program crash in development, so we can [`.expect()`](https://doc.rust-lang.org/std/result/enum.Result.html#method.unwrap) the `Result` here, which will return the parsed `Color` for us, or crash if its malformatted. The message we write here is written in the format [suggested in the docs](https://doc.rust-lang.org/std/result/enum.Result.html#recommended-message-style) for `.expect`.

A hex code that doesn't parse will panic and crash our program with a message like this:

```
❯ cargo run
   Compiling boxes v0.1.0 (/rust-adventure/2048)
    Finished dev [unoptimized + debuginfo] target(s) in 1.13s
     Running `target/debug/boxes`
thread 'main' panicked at src/main.rs:8:36:
developer should have provided a valid hex code: Length
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```

For each type like `Color`, there can only be one `Resource` value with that type. That is: `Resource`s are singletons. This is why we wrap the `Color` in the `ClearColor` struct. Bevy will look for the `ClearColor` resource and use it if we specify it.

When we `cargo run` now, we see the board in the middle of a blueish background.

![a board with a clear color](https://res.cloudinary.com/dilgcuzda/image/upload/v1713942738/workshops/2048-with-bevy-ecs/bevy-0.13/007-clear-color.avif)